+++ packages/core/src/graph/graph.ts
@@ -9,7 +9,7 @@ import {
   floydWarshall,
 } from '@antv/algorithm';
 import { IAbstractGraph } from '../interface/graph';
-import { IEdge, INode, ICombo, IItemBaseConfig } from '../interface/item';
+import { IEdge, INode, ICombo } from '../interface/item';
 import {
   GraphAnimateConfig,
   GraphOptions,
@@ -340,6 +340,9 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
 
       // 存储图上的 tooltip dom，方便销毁
       tooltips: [],
+
+      // 达到这一节点数量(默认值 1000)，将开启性能优化模式。目前包括：节点状态样式变更是否影响相关边的更新
+      optimizeThreshold: 1000
     };
   }
 
@@ -1748,6 +1751,8 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
    * @param children 添加到 Combo 中的元素，包括节点和 combo
    */
   public createCombo(combo: string | ComboConfig, children: string[]): void {
+    const itemController: ItemController = this.get('itemController');
+
     this.set('comboSorted', false);
     // step 1: 创建新的 Combo
     let comboId = '';
@@ -1767,14 +1772,52 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
       comboConfig = combo;
     }
 
-    // step2: 更新 children，根据类型添加 comboId 或 parentId
+    // step 2: Pull children out of their parents
+    let comboTrees = this.get('comboTrees');
+    const childrenIdsSet = new Set<string>(children);
+    const pulledComboTreesById = new Map<string, ComboTree>();
+
+    if (comboTrees) {
+      comboTrees.forEach(ctree => {
+        traverseTreeUp<ComboTree>(ctree, (treeNode, parentTreeNode, index) => {
+          if (childrenIdsSet.has(treeNode.id)) {
+            if (parentTreeNode) {
+              const parentItem = this.findById(parentTreeNode.id) as ICombo;
+              const item = this.findById(treeNode.id) as INode | ICombo;
+
+              // Removing current item from the tree during the traversal is ok because children traversal is done
+              // in an *inverse order* - indices of the next-traversed items are not disturbed by the removal.
+              parentTreeNode.children.splice(index, 1);
+              parentItem.removeChild(item);
+
+              // We have to update the parent node geometry since nodes were removed from them, _while they are still visible_
+              // (combos may be moved inside the new combo and become hidden)
+              itemController.updateCombo(parentItem, parentTreeNode.children);
+            }
+
+            if (treeNode.itemType === 'combo') {
+              pulledComboTreesById.set(treeNode.id, treeNode);
+            }
+          }
+
+          return true;
+        });
+      });
+      comboTrees = comboTrees.filter(ctree => !childrenIdsSet.has(ctree.id));
+
+      this.set('comboTrees', comboTrees);
+    }
+
+    // step 3: 更新 children，根据类型添加 comboId 或 parentId
     const trees: ComboTree[] = children.map(elementId => {
       const item = this.findById(elementId);
       const model = item.getModel();
 
       let type = '';
       if (item.getType) type = item.getType();
-      const cItem: ComboTree = {
+
+      // Combos will be just moved around, so their children can be preserved
+      const cItem: ComboTree = pulledComboTreesById.get(elementId) || {
         id: item.getID(),
         itemType: type as 'node' | 'combo',
       };
@@ -1792,23 +1835,25 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
 
     comboConfig.children = trees;
 
-    // step 3: 添加 Combo，addItem 时会将子将元素添加到 Combo 中
+    // step 4: 添加 Combo，addItem 时会将子将元素添加到 Combo 中
     this.addItem('combo', comboConfig, false);
     this.set('comboSorted', false);
 
-    // step4: 更新 comboTrees 结构
-    const comboTrees = this.get('comboTrees');
-    (comboTrees || []).forEach(ctree => {
-      traverseTreeUp<ComboTree>(ctree, child => {
-        if (child.id === comboId) {
-          child.itemType = 'combo';
-          child.children = trees as ComboTree[];
-          return false;
-        }
-        return true;
-      });
-    });
+    // step 5: 更新 comboTrees 结构
     if (comboTrees) {
+      comboTrees.forEach(ctree => {
+        traverseTree<ComboTree>(ctree, (treeNode) => {
+          // Set the children to the newly created combo
+          if (treeNode.id === comboId) {
+            treeNode.itemType = 'combo';
+            treeNode.children = trees as ComboTree[];
+            return false;
+          }
+
+          return true;
+        });
+      });
+
       this.sortCombos();
     }
   }