+++ packages/core/tests/unit/graph/graph-combo-spec.ts
@@ -1,6 +1,5 @@
 import Graph from '../implement-graph';
-import { ICombo } from '../../../src/interface/item';
-import { GraphData } from '../../../src/types';
+import { GraphData, ICombo, ComboTree, Util } from '../../../src';
 import { clone } from '@antv/util';
 
 const div = document.createElement('div');
@@ -113,13 +112,88 @@ const data = {
 };
 
 describe('graph with combo', () => {
-  it('uncombo', () => {
-    const graph = new Graph({
+  let graph: Graph;
+
+  function makeGraph(config = {}) {
+    graph = new Graph({
       container: div,
       width: 500,
       height: 600,
+      ...config
     });
     graph.read(clone(data));
+  }
+
+  it('createCombo', () => {
+    makeGraph();
+
+    graph.addItem('node', { id: 'top-level-node' });
+
+    const newComboChildIds = [
+      'top-level-node', // a top-level node
+      '0', // Inside top-level combo 'a'
+      '2', // Inside top-level combo 'b'
+      '5', // Inside depth-1 combo 'e'
+      'a', // Top-level combo
+      'e', // Depth-1 combo
+    ];
+
+    graph.createCombo('new-combo', newComboChildIds);
+
+    const newCombo = graph.findById('new-combo') as ICombo;
+
+    // Children are assigned correctly to the new combo parent
+    newComboChildIds.forEach(id => {
+      const itemModel = graph.findById(id).getModel();
+      expect(itemModel.comboId || itemModel.parentId).toBe('new-combo');
+    });
+
+    // Combo children are updated correctly
+    const expectedChildrenById = {
+      a: ['1'], // '0' has been transferred to new combo
+      b: ['3'], // 'e' moved to new combo
+      c: ['4'], // Not changed
+      e: ['6'], // '5' moved to new combo
+      'new-combo': newComboChildIds,
+    }
+
+    Object.entries(expectedChildrenById).forEach(([comboId, expectedChildIds]) => {
+      const combo = graph.findById(comboId) as ICombo;
+      const children = [...combo.getNodes(), ...combo.getCombos()];
+      const childIds = children.map(item => item.getID());
+
+      // Double-contains because we don't care about order
+      expect(childIds).toEqual(expect.arrayContaining(expectedChildIds));
+      expect(expectedChildIds).toEqual(expect.arrayContaining(childIds));
+    });
+
+    // Children should have been removed from their original parents in the comboTrees
+    const expectedTopLevelComboIds = ['b', 'c', 'new-combo'];
+
+    const comboTrees = graph.get('comboTrees');
+    const topLevelComboIds = comboTrees.map(comboTree => comboTree.id);
+
+    expect(topLevelComboIds).toEqual(expect.arrayContaining(expectedTopLevelComboIds));
+    expect(expectedTopLevelComboIds).toEqual(expect.arrayContaining(topLevelComboIds));
+
+    comboTrees.forEach(ctree => {
+      Util.traverseTree<ComboTree>(ctree, (node: ComboTree): boolean => {
+        if (node.itemType === 'combo') {
+          const childIds = node.children.map(child => child.id);
+          const expectedChildIds = expectedChildrenById[node.id];
+
+          expect(childIds).toEqual(expect.arrayContaining(expectedChildIds));
+          expect(expectedChildIds).toEqual(expect.arrayContaining(childIds));
+        }
+
+        return true;
+      });
+    });
+  });
+
+  it('uncombo', () => {
+    makeGraph();
+
     // uncombo with combo item
     graph.uncombo(graph.findById('a') as ICombo);
     expect(graph.getCombos().length).toBe(3);
@@ -139,12 +213,8 @@ describe('graph with combo', () => {
     graph.destroy();
   });
   it('uncombo subcombo', () => {
-    const graph = new Graph({
-      container: div,
-      width: 500,
-      height: 600,
-    });
-    graph.read(clone(data));
+    makeGraph();
+
     // uncombo the parent first and then uncombo the subcombo
     graph.uncombo(graph.findById('b') as ICombo);
     expect(graph.getCombos().length).toBe(3);
@@ -156,12 +226,8 @@ describe('graph with combo', () => {
   });
 
   it('get combo children', () => {
-    const graph = new Graph({
-      container: div,
-      width: 500,
-      height: 600,
-    });
-    graph.read(clone(data));
+    makeGraph();
+
     const comboB: ICombo = graph.findById('b') as ICombo;
     const comboBChildren = comboB.getChildren();
     expect(comboBChildren.nodes.length).toBe(2);
@@ -171,12 +237,8 @@ describe('graph with combo', () => {
   });
 
   it('hide and show a combo', () => {
-    const graph = new Graph({
-      container: div,
-      width: 500,
-      height: 600,
-    });
-    graph.read(clone(data));
+    makeGraph();
+
     // hide an item itself
     const comboA: ICombo = graph.findById('a') as ICombo;
     comboA.hide();
@@ -208,10 +270,7 @@ describe('graph with combo', () => {
   });
 
   it('collapse expand combo', () => {
-    const graph = new Graph({
-      container: div,
-      width: 500,
-      height: 600,
+    makeGraph({
       defaultCombo: {
         size: 10,
         padding: 1,
@@ -223,7 +282,7 @@ describe('graph with combo', () => {
       },
       animate: true,
     });
-    graph.read(clone(data));
+
     // collapse / expand the combo with invalid id
     graph.collapseCombo('invalidid');
     graph.expandCombo('invalidid');
@@ -271,12 +330,8 @@ describe('graph with combo', () => {
   });
 
   it('combo mapper', () => {
-    const graph = new Graph({
-      container: div,
-      width: 500,
-      height: 600,
-      groupByTypes: false,
-    });
+    makeGraph({ groupByTypes: false });
+
     graph.combo(() => {
       return {
         style: {
@@ -294,12 +349,7 @@ describe('graph with combo', () => {
   });
 
   it('getComboChildren', () => {
-    const graph = new Graph({
-      container: div,
-      width: 500,
-      height: 600,
-    });
-    graph.read(clone(data));
+    makeGraph();
 
     // getComboChildren with id
     const childrenById = graph.getComboChildren('b');
@@ -318,10 +368,7 @@ describe('graph with combo', () => {
   });
 
   it('changeData with combos', () => {
-    const graph = new Graph({
-      container: div,
-      width: 500,
-      height: 600,
+    makeGraph({
       layout: {
         type: 'comboForce',
       },
@@ -331,7 +378,7 @@ describe('graph with combo', () => {
       fitView: true,
       groupByTypes: false,
     });
-    graph.read(clone(data));
+
     const newData = {
       nodes: [
         {
@@ -405,17 +452,15 @@ describe('graph with combo', () => {
     // expect(comboAChildren.combos.length).toBe(2);
     // graph.destroy();
   });
+
   it('updateComboTree', () => {
-    const graph = new Graph({
-      container: div,
-      width: 500,
-      height: 600,
+    makeGraph({
       layout: {
         type: 'comboForce',
       },
       fitView: true,
     });
-    graph.read(clone(data));
+
     graph.updateComboTree('4', 'b');
     graph.layout();
     expect(graph.get('comboTrees').length).toBe(3);
@@ -433,10 +478,7 @@ describe('graph with combo', () => {
     graph.destroy();
   });
   it('add combo item', () => {
-    const graph = new Graph({
-      container: div,
-      width: 500,
-      height: 600,
+    makeGraph({
       layout: {
         type: 'comboForce',
       },
@@ -445,11 +487,11 @@ describe('graph with combo', () => {
       },
       fitView: true,
     });
-    graph.read(clone(data));
+
     const newComboItem = graph.addItem('combo', {
       id: 'new combo',
       label: 'new combo',
-    });
+    }) as ICombo;
     expect(newComboItem.getChildren().nodes.length).toBe(0);
     expect(newComboItem.getChildren().combos.length).toBe(0);
     expect(graph.getCombos().length).toBe(5);
@@ -459,7 +501,7 @@ describe('graph with combo', () => {
       id: 'new combo 2',
       label: 'new combo 2',
       parentId: 'new combo',
-    });
+    }) as ICombo;
     graph.addItem('combo', {
       id: 'new combo 3',
       label: 'new combo 3',
@@ -501,10 +543,7 @@ describe('graph with combo', () => {
   });
 
   it('graph.save with combos', () => {
-    const graph = new Graph({
-      container: div,
-      width: 500,
-      height: 600,
+    makeGraph({
       layout: {
         type: 'comboForce',
       },
@@ -513,7 +552,7 @@ describe('graph with combo', () => {
       },
       animate: true,
     });
-    graph.read(clone(data));
+
     const result = graph.save() as GraphData;
     expect(result.combos).not.toBe(undefined);
     expect(result.combos.length).toBe(4);